+++ lib/jscalendar.ex
@@ -130,7 +130,7 @@ if Code.ensure_loaded?(JSCalendar) do
         iex> {:ok, set} = Tempo.JSCalendar.from_jscalendar(json)
         iex> [interval] = Tempo.IntervalSet.to_list(set)
         iex> Tempo.to_iso8601(interval)
-        "2026Y6M2DT9H0M0S/2026Y6M2DT10H0M0S"
+        "2026Y6M2DT9H0M0S/T10H0M0S"
         iex> Tempo.Interval.metadata(interval).title
         "Quarterly review"